+++ packages/config/server.ts
@@ -18,3 +18,4 @@ export * from "./production-launch";
 export * from "./production-load";
 export * from "./storage-connect-origin";
 export * from "./workflows";
+export * from "./kie-image-certification";